+++ crypto/build.info
@@ -73,7 +73,7 @@ $UTIL_COMMON=\
 $UTIL_DEFINE=$CPUIDDEF
 
 SOURCE[../libcrypto]=$UTIL_COMMON \
-        mem.c mem_sec.c mem_str.c mem_dbg.c \
+        mem.c mem_sec.c mem_dbg.c \
         cversion.c info.c cpt_err.c ebcdic.c uid.c o_time.c o_dir.c \
         o_fopen.c getenv.c o_init.c o_fips.c init.c trace.c provider.c \
         asn1_dsa.c packet.c $UPLINKSRC
